BUGFIX: form binding after errors
If an update form was submitted with errors, the entity id was not passed back. As a result the ui button changed from "Update" to "Create".

This was a result of removing Pet Clinic's controller wide @ModelAttribute. This was removed to handle Stakeholders having the same attribute names as Specifications. Using the controller wide binder caused both values to be overwritten.